Begin refactoring to decouple components
The connection objects call the server object and vice versa. This
refactor begins to reverse that, making it so that the connection
object has no knowledge of the server object.

We're also deferring all of the connection.write calls out to the
connection object, so that the server object doesn't need to be
aware of the protocol.