Start sending WORK_FAIL packets next to WORK_EXCEPTION packets
The protocol spec claims that WORK_EXCEPTION packets should be fatal, but
the actual implementation in ALL server versions does not actually do this.

The c++ version looks like it should (to my non-c++-ey eyes) so perhaps it has
a bug lurking somewhere.

Regardless, in order to work with all extant servers we _either_ need to expose
WORK_EXCEPTION like we do WORK_WARNING, OR we need to send WORK_FAIL with it.
We're doing the latter.

Fixes #15, for realsie.